it is an error not to fill defaultExts
Its type makes it look like [] is a good default, but this field is
normally filled with information about a long list of extensions, and so
it is a good thing that the "error" call is there to remind us to fill
that field properly. Of course, it would be even better if the code was
restructured in a way which statically guarantees that the field is
initialiaed.