remove spurious ghc upper bound
Each time we suppport a new version of ghc, it takes us a while to
figure out why the tests are failing, until we notice that upper bound
and bump it up. I see no reason for a restrictive upper bound here, so
let's just remove the upper bound to make supporting new versions of ghc
easier.