don't check the readme as a unit test
checking the readme as part of the unit tests was a bad idea because the
test relies on doctest, which only installs and runs cleanly  on some of
the configurations under which we want the unit tests to pass. whereas
for the readme, there's no need to check that its code runs fine under
many different configurations.