delay computation of 'runInterp'
An IOTestCase is allowed to run IO before calling 'runInterpreter', and
to instead call a variant of 'runInterpreter' such as
'runInterpreterWithArgs', as long as it threads it through the
IOTestCase constructor. Unfortunately, this threading API was limiting:
only a single variant per test could be used, and IO could not be used
to construct the variant. This new API fixes both issues.