test that hint respects ghc environment files
A very common use case for hint is to interpret code which has access to
the same packages as the compiled program which calls hint. the easiest
way to accomplish that is to compile that program using
"cabal build --write-ghc-environment-files=always", as that will write
down all the necessary information to a file which hint can read at
runtime. This test makes sure that this common workflow continues to
work in the future.