simplify installSupportModule
Previously, we were adding the phantom module to the list of targets,
and then loading all the targets, and then if there are more targets to
load, we were setting the targets to those plus the phantom module, and
then loading all the targets again.

Now, we simply set the targets to those plus the phantom module, and
then we load all the targets, once.