always call LoadAllTargets
In the previous two commits, I said that installSupportModule was
calling LoadAllTargets twice in a row. That was only true on ghc-9.4+;
on older versions of ghc, the first call was actually a call to
"LoadUpTo phantomModule".

Now that there is only one call, "LoadUpTo phantomModule" is
insufficient, because we want to load all the targets, not just the
phantom module.